Add window.confirm when deleting beverage entries

+++ client/src/components/EditableDataTable/EditableDataTable.js
@@ -65,8 +65,8 @@ class EditableDataTable extends Component {
       newData: {},
       isAvailable: null,
       isAdmin: null,
-      order: 'asc',
-      orderBy: 'firstName',
+      order: "asc",
+      orderBy: "firstName",
       selected: null
     };
   }
@@ -114,21 +114,20 @@ class EditableDataTable extends Component {
     // let newSelected = [];
 
     // if (selectedIndex === -1) {
-  //     newSelected = newSelected.concat(selected, id);
-  //   } else if (selectedIndex === 0) {
-  //     newSelected = newSelected.concat(selected.slice(1));
-  //   } else if (selectedIndex === selected.length - 1) {
-  //     newSelected = newSelected.concat(selected.slice(0, -1));
-  //   } else if (selectedIndex > 0) {
-  //     newSelected = newSelected.concat(
-  //       selected.slice(0, selectedIndex),
-  //       selected.slice(selectedIndex + 1),
-  //     );
-  //   }
+    //     newSelected = newSelected.concat(selected, id);
+    //   } else if (selectedIndex === 0) {
+    //     newSelected = newSelected.concat(selected.slice(1));
+    //   } else if (selectedIndex === selected.length - 1) {
+    //     newSelected = newSelected.concat(selected.slice(0, -1));
+    //   } else if (selectedIndex > 0) {
+    //     newSelected = newSelected.concat(
+    //       selected.slice(0, selectedIndex),
+    //       selected.slice(selectedIndex + 1),
+    //     );
+    //   }
     if (selected === id) {
       this.setState({ selected: null });
-    } 
-    else {
+    } else {
       this.setState({ selected: id });
     }
   };
@@ -197,37 +196,40 @@ class EditableDataTable extends Component {
   //   });
   // };
 
-
   deleteBeverage = bevId => {
-    API.deleteBeverage(bevId)
-    .then((resp) => {
-      if (resp.status === 200 && resp.statusText === "OK") {
-        console.log(`All good. Deleted ${resp.data} beverage!`);
-        this.getAllBeverages();
-      }
-      else {
-        console.log(
-          `Delete beverage returned non-error status code: please debug`
-        );
-        if (resp.status === 500) {
-          // Error: Request failed with status code 500
-          console.log(`500 error`);
-        }
-        console.log(resp.status, resp.statusText);
-      }
-    })
-    .catch(err => {
-      console.log(`Error updating beverage: ${err}`);
-      if (err.Error === "Network Error") {
-        console.log(
-          `Internet disconnected, undo state changes and fire modal to let user know`
-        );
-        // this.rollbackStateAfterAPIFail(bevObj);
-      } else {
-        console.log(`Non network-related error. Please debug: ${err}`);
-      }
-    });
-  }
+    const del = window.confirm(
+      "Are you sure you want to delete the selected beverage?"
+    );
+    if (del) {
+      API.deleteBeverage(bevId)
+        .then(resp => {
+          if (resp.status === 200 && resp.statusText === "OK") {
+            console.log(`All good. Deleted ${resp.data} beverage!`);
+            this.getAllBeverages();
+          } else {
+            console.log(
+              `Delete beverage returned non-error status code: please debug`
+            );
+            if (resp.status === 500) {
+              // Error: Request failed with status code 500
+              console.log(`500 error`);
+            }
+            console.log(resp.status, resp.statusText);
+          }
+        })
+        .catch(err => {
+          console.log(`Error updating beverage: ${err}`);
+          if (err.Error === "Network Error") {
+            console.log(
+              `Internet disconnected, undo state changes and fire modal to let user know`
+            );
+            // this.rollbackStateAfterAPIFail(bevObj);
+          } else {
+            console.log(`Non network-related error. Please debug: ${err}`);
+          }
+        });
+    }
+  };
   // Handle updating database when a beverage is changed
   sendUpdateBeverage = bevObj => {
     API.changeBeverage(bevObj)
@@ -353,7 +355,7 @@ class EditableDataTable extends Component {
         <Paper className="overflow-table">
           <Table>
             {this.props.type === "bevs" ? (
-              <BevTableHeader 
+              <BevTableHeader
                 deleteBev={this.deleteBeverage}
                 isSelected={this.state.selected}
                 order={order}
@@ -368,12 +370,12 @@ class EditableDataTable extends Component {
               />
             )}
             <TableBody>
-              {this.props.type === "bevs" ? 
-                this.state.data[0] !== "Loading..." &&
-                  stableSort(data, getSorting(order, orderBy))
-                    .map(row => {
-                      const isSelected = this.isSelected(row._id);
-                      return (<BevTableRow
+              {this.props.type === "bevs"
+                ? this.state.data[0] !== "Loading..." &&
+                  stableSort(data, getSorting(order, orderBy)).map(row => {
+                    const isSelected = this.isSelected(row._id);
+                    return (
+                      <BevTableRow
                         key={row._id}
                         handleClick={this.handleClick}
                         handleFieldChange={this.handleFieldChange}
@@ -410,9 +412,13 @@ class EditableDataTable extends Component {
             <p className="text-center">No beverages! Try creating some.</p>
           )}
         </Paper>
-        {this.props.type === "users" ? <Paper><AddUser /></Paper> : null}
+        {this.props.type === "users" ? (
+          <Paper>
+            <AddUser />
+          </Paper>
+        ) : null}
         {this.props.type === "bevs" && this.state.data[0] !== "Loading..." && (
-            <SaveButton saveHandler={this.saveHandler} />
+          <SaveButton saveHandler={this.saveHandler} />
         )}
       </>
     );
